login,signup finish

+++ crud/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.utils import timezone
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
+from django.core.paginator import Paginator
 
 from .models import *
 from .forms import CrudUpdate, InfoUpdate, QnaUpdate
@@ -22,7 +23,10 @@ def index(request):
 # 자유게시판~~
 def free(request):
     cruds = Crud.objects.all().order_by('-pub_date')
-    return render(request, 'free.html', {'cruds':cruds})
+    paginator = Paginator(cruds, 5) # 페이지네이션 구현
+    page = request.GET.get("page")
+    posts = paginator.get_page(page)
+    return render(request, 'free.html', {'cruds' : posts})
     
 def detail(request, crud_id):
     crud_detail = get_object_or_404(Crud, pk=crud_id)
@@ -46,35 +50,38 @@ def new(request):
     return render(request, 'new.html')
 
 def create(request):
-    crud = Crud()
-    crud.title = request.GET['title']
-    crud.body = request.GET['body']
-    crud.name = request.GET['name']
-    crud.pub_date = timezone.datetime.now()
-    crud.save()
-    return redirect('/crud/' + str(crud.id))
+    if 'name' in request.session:
+        crud = Crud()
+        crud.title = request.GET['title']
+        crud.body = request.GET['body']
+        crud.name = request.GET['name']
+        crud.pub_date = timezone.datetime.now()
+        crud.save()
+        return redirect('/crud/' + str(crud.id))
 
 def delete(request, crud_id):
-    crud = Crud.objects.get(id=crud_id)
-    crud.delete()
-    return redirect('/free/')
+    if 'name' in request.session:
+        crud = Crud.objects.get(id=crud_id)
+        crud.delete()
+        return redirect('/free/')
 
 def edit(request, crud_id):
-    crud = Crud.objects.get(id=crud_id)
+    if 'name' in request.session:
+        crud = Crud.objects.get(id=crud_id)
 
-    if request.method =='POST':
-        form = CrudUpdate(request.POST)
-        if form.is_valid():
-            crud.title = form.cleaned_data['title']
-            crud.body = form.cleaned_data['body']
-            crud.name = request.GET['name']
-            crud.pub_date=timezone.now()
-            crud.save()
-            return redirect('/detail/' + str(crud.id))
-    else:
-        form = CrudUpdate(instance = crud)
- 
-        return render(request,'edit.html', {'form':form})
+        if request.method =='POST':
+            form = CrudUpdate(request.POST)
+            if form.is_valid():
+                crud.title = form.cleaned_data['title']
+                crud.body = form.cleaned_data['body']
+                crud.name = request.GET['name']
+                crud.pub_date=timezone.now()
+                crud.save()
+                return redirect('/detail/' + str(crud.id))
+        else:
+            form = CrudUpdate(instance = crud)
+    
+            return render(request,'edit.html', {'form':form})
 
 def crudComment(request): # 댓글 생성
     if request.method == "POST":
@@ -94,7 +101,10 @@ def crudComment(request): # 댓글 생성
 # 정보게시판~~
 def infoList(request):
     infos = Info.objects.all().order_by('-pub_date')
-    return render(request, 'infoList.html', {'infos':infos})
+    paginator = Paginator(infos, 5) # 페이지네이션 구현
+    page = request.GET.get("page")
+    posts = paginator.get_page(page)
+    return render(request, 'infoList.html', {'infos':posts})
 
 def infoDetail(request, info_id):
     info_detail = get_object_or_404(Info, pk=info_id)
@@ -165,7 +175,10 @@ def infoComment(request): # 댓글 생성
 # 질문 게시판~~
 def qnaList(request):
     qnas = Qna.objects.all().order_by('-pub_date')
-    return render(request,'qnaList.html', {'qnas': qnas})
+    paginator = Paginator(qnas, 5) # 페이지네이션 구현
+    page = request.GET.get("page")
+    posts = paginator.get_page(page)
+    return render(request,'qnaList.html', {'qnas': posts})
 
 def qnaDetail(request, qna_id):
     qna_detail = get_object_or_404(Qna, pk=qna_id)
@@ -186,35 +199,38 @@ def qnaDetail(request, qna_id):
     return render(request, 'qnaDetail.html' , content)
 
 def qnaCreate(request):
-    qna = Qna()
-    qna.title = request.GET['title']
-    qna.body = request.GET['body']
-    qna.name = request.GET['name']
-    qna.pub_date = timezone.datetime.now()
-    qna.save()
-    return redirect('/qna/' + str(qna.id))
+    if 'name' in request.session:
+        qna = Qna()
+        qna.title = request.GET['title']
+        qna.body = request.GET['body']
+        qna.name = request.GET['name']
+        qna.pub_date = timezone.datetime.now()
+        qna.save()
+        return redirect('/qna/' + str(qna.id))
 
 def qnaDelete(request, qna_id):
-    qna = Qna.objects.get(id=qna_id)
-    qna.delete()
-    return redirect('/qnaList/')
+    if 'name' in request.session:
+        qna = Qna.objects.get(id=qna_id)
+        qna.delete()
+        return redirect('/qnaList/')
 
 def qnaEdit(request, qna_id):
-    qna = Qna.objects.get(id=qna_id)
+    if 'name' in request.session:
+        qna = Qna.objects.get(id=qna_id)
 
-    if request.method =='POST':
-        form = QnaUpdate(request.POST)
-        if form.is_valid():
-            qna.title = form.cleaned_data['title']
-            qna.body = form.cleaned_data['body']
-            qna.name = request.GET['name']
-            qna.pub_date=timezone.now()
-            qna.save()
-            return redirect('/qnaDetail/' + str(qna.id))
-    else:
-        form = QnaUpdate(instance = qna)
- 
-        return render(request,'qnaEdit.html', {'form':form})
+        if request.method =='POST':
+            form = QnaUpdate(request.POST)
+            if form.is_valid():
+                qna.title = form.cleaned_data['title']
+                qna.body = form.cleaned_data['body']
+                qna.name = request.GET['name']
+                qna.pub_date=timezone.now()
+                qna.save()
+                return redirect('/qnaDetail/' + str(qna.id))
+        else:
+            form = QnaUpdate(instance = qna)
+    
+            return render(request,'qnaEdit.html', {'form':form})
     
 def qnaNew(request):
     return render(request, 'qnaNew.html')
